bigquery/requirements.txt
authlib
fastapi
httpx
itsdangerous
jinja2
python-jose
python-multipart
uvicorn
